+++ src/Engine/Physics/PhysX/Actors/LinkActor.cpp
@@ -25,8 +25,8 @@ namespace owds::physx {
   LinkActor::~LinkActor() noexcept
   {}
 
-  void LinkActor::setup(const std::array<float, 3>& position,
-                        const std::array<float, 4>& orientation)
+  void LinkActor::setup(const std::array<double, 3>& position,
+                        const std::array<double, 4>& orientation)
   {
     (void)position;
     (void)orientation;
@@ -69,7 +69,7 @@ namespace owds::physx {
     px_link_->setMass(static_cast<::physx::PxReal>(mass_kg));
   }
 
-  void LinkActor::setPositionAndOrientation(const std::array<float, 3>& position, const std::array<float, 4>& orientation)
+  void LinkActor::setPositionAndOrientation(const std::array<double, 3>& position, const std::array<double, 4>& orientation)
   {
     (void)position;
     (void)orientation;
